requirements.txt
python-telegram-bot>=20.0
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
feedparser>=6.0.0



